% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_lottie.R
\name{include_lottie}
\alias{include_lottie}
\title{Include Lottie Functionality within Shiny}
\usage{
include_lottie(version = "5.12.2")
}
\arguments{
\item{version}{A character string specifying the version of the Lottie library to source via CDN.}
}
\value{
A list of HTML tags to be included within the head element of a Shiny application.
}
\description{
Responsible for retrieving the Lottie library and initialising the relevant shinyLottie JavaScript code.
As such, this function \strong{must} be included within the UI object of a \code{\link[shiny:shinyApp]{shinyApp}} in order to enable shinyLottie functionality.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
library(shinyLottie)

ui <- fluidPage(
  include_lottie(),
  lottie_animation(
    path = "shinyLottie/example.json",
    name = "my_animation"
  )
)

server <- function(input, output, session) {}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
