% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lottie_animation_methods.R
\name{lottie_animation_methods}
\alias{lottie_animation_methods}
\title{'Lottie' Animation Methods}
\description{
These functions provide methods for modifying the playback options of existing 'Lottie' animations.
}
\details{
\itemize{
  \item \code{\link{lottie_setSpeed}}: Set the speed of the animation.
  \item \code{\link{lottie_setDirection}}: Set the direction of the animation.
  \item \code{\link{lottie_setLoop}}: Set whether the animation should loop.
  \item \code{\link{lottie_goToAndStop}}: Move to a specific frame or time and stop.
  \item \code{\link{lottie_goToAndPlay}}: Move to a specific frame or time and play.
  \item \code{\link{lottie_playSegments}}: Play specific segments of the animation.
  \item \code{\link{lottie_setSubframe}}: Set whether to use subframes when rendering the animation.
  \item \code{\link{lottie_destroy}}: Destroy the specified animation instance.
}
}
